+++ tests/testthat/test-get.R
@@ -3,6 +3,6 @@ test_that("functions to get elevations from sf objects work", {
     skip(message = "Skipping test, MAPBOX token in .Renviron needed")
   r = cyclestreets_route
   e = elevation_get(r)
-  expect_equal(class(e)[1], "RasterLayer")
+  expect_true(methods::is(e, "SpatRaster"))
 })